refactor(cli): minor refactor moving cli.rs helper function into a separate mod plus additional documentation to README
