+++ README.md
@@ -118,23 +118,37 @@ $ SAFE_AUTH_SECRET="<secret>" SAFE_AUTH_PASSWORD="<password>" cargo run -- --pre
 Logged in the SAFE Network successfully!
 ```
 
-Do note, that _both_ the secret and password env vars must be set to use this auth method. If only one is set, an error will be thrown.
+Do note, that _both_ the secret and password environment variables must be set to use this method. If only one is set, an error will be thrown.
 
 ### Authorising an application
 ```
 $ cargo run -- --pretty --req <auth req string>
 Secret:
 Password:
+Logged in the SAFE Network successfully!
+The following application authorisation request was received:
++---------------------------------+--------------+------------------+----------------------------------------------+
+| Id                              | Name         | Vendor           | Permissions requested                        |
++---------------------------------+--------------+------------------+----------------------------------------------+
+| <app ID>                        | <app's name> | <vendor name>    | Own container: false                         |
+|                                 |              |                  | Default containers: <permissions requested>  |
++---------------------------------+--------------+------------------+----------------------------------------------+
+Allow authorisation? [y/N]:
+Authorisation will be allowed...
 Authorisation response string: <auth response>
 ```
 
-
-As an example, the following command passes a valid encoded authorisation request as the value of the `--req` argument:
+As you can see beafore each authorisation request is allowed, the user is prompted for confirm. Optionally, this prompt can be disabled to have the `safe_auth` to automatically allow all incoming authorisation requests. We can do this by passing the `--allow-all-auth` argument in the command line:
 ```
-$ cargo run -- --pretty --req bAAAAAAFBMHKYWAAAAAABWAAAAAAAAAAANZSXILTNMFUWI43BMZSS45DFON2C453FMJQXA4BONFSAACYAAAAAAAAAABLWKYSBOBYCAVDFON2A2AAAAAAAAAAAJVQWSZCTMFTGKICMORSC4AACAAAAAAAAAAAAUAAAAAAAAAAAL5SG6Y3VNVSW45DTAEAAAAAAAAAAAAIAAAAAOAAAAAAAAAAAL5YHKYTMNFRQCAAAAAAAAAAAAAAAAAAB
+$ cargo run -- --pretty --allow-all-auth --req <auth req string>
 ```
 
-The expected encoded authorisation request is the one that can be generated by an application using the SAFE API, e.g. an application using the `safe_app_nodejs` would make use of the [genAuthUri](https://docs.maidsafe.net/safe_app_nodejs/authinterface#genAuthUri), [genConnUri](https://docs.maidsafe.net/safe_app_nodejs/authinterface#genConnUri), [genContainerAuthUri](https://docs.maidsafe.net/safe_app_nodejs/authinterface#genContainerAuthUri), or [genShareMDataUri](https://docs.maidsafe.net/safe_app_nodejs/authinterface#genShareMDataUri) functions to generate such encoded string.
+As an example, the following command passes a valid encoded authorisation request as the value of the `--req` argument and it allows the autorisation to be made without prompting:
+```
+$ cargo run -- --pretty --allow-all-auth --req bAAAAAABU6IEAEAAAAAACMAAAAAAAAAAANZSXILTNMFUWI43BMZSS45DFON2C4YLVORUGK3TUNFRWC5DPOIXGG3DJFZUWIAILAAAAAAAAAAAF65DFON2F643DN5YGKGYAAAAAAAAAABJHK43UEBAXK5DIMVXHI2LDMF2G64RAINGESICUMVZXIEAAAAAAAAAAABGWC2LEKNQWMZJONZSXIICMORSAAAIAAAAAAAAAAADQAAAAAAAAAAC7OB2WE3DJMMAQAAAAAAAAAAAAAAAAAAI
+```
+
+The expected encoded authorisation request string is the one that can be generated by any application using the SAFE API, e.g. an application using the `safe_app_nodejs` would make use of the [genAuthUri](https://docs.maidsafe.net/safe_app_nodejs/authinterface#genAuthUri), [genConnUri](https://docs.maidsafe.net/safe_app_nodejs/authinterface#genConnUri), [genContainerAuthUri](https://docs.maidsafe.net/safe_app_nodejs/authinterface#genContainerAuthUri), or [genShareMDataUri](https://docs.maidsafe.net/safe_app_nodejs/authinterface#genShareMDataUri) functions to generate such encoded string.
 
 The output obtained from the `safe_auth` CLI command when passing a `--req` argument, can then be used by such a Nodejs application to connect to the SAFE Network with the [loginFromUri](https://docs.maidsafe.net/safe_app_nodejs/authinterface#loginFromUri) function.
 
@@ -201,7 +215,7 @@ Exposing service on 127.0.0.1:41805
 
 Then on a separate terminal you can authorise an application with:
 ```
-$ curl -X POST http://localhost:41805/authorise/<auth req string>
+$ curl -X GET http://localhost:41805/authorise/<auth req string>
 ```
 
 ## License
